end product api

+++ controller/product.js
@@ -1,19 +1,25 @@
 const Product = require('../model/product');
 const slugify = require('slugify')
-const validationResult = require('express-validator').validationResult;
 const fs = require('fs')
+const Category = require('../model/category');
+const mongoose = require('mongoose');
 let uploadsPath =  __dirname.replace("controller", "uploads") + '\\' ;
 
 exports.getAllproducts = async (req, res, next) => {
     const query = Product.find()
     query.select("_id name category image")
     const products = await query.exec();
-    products.forEach(product => {
-        product.show = {
-            method: "GET",
-            url: process.env.URL + "api/products/" + product.id
+    products = products.map(product => {
+        return {
+            id: product._id,
+            name: product.name,
+            category: product.category,
+            image: process.env.URL + product.image,
+            productPage: {
+                method: "GET",
+                url: process.env.URL + "api/products/" + product.id
+            },
         }
-        product.image = process.env.URL+ product.image
     })
     if (products.length > 0)
         return res.status(200).json({products})
@@ -24,7 +30,6 @@ exports.getAllproducts = async (req, res, next) => {
 }
 exports.getProduct = async (req, res, next) => {
     const id = req.params.productId;
-    console.log(id)
     const product = await Product.findById(id);
     if (!product) {
         return res.status(500).json({message: "something went wrong"})
@@ -56,84 +61,33 @@ exports.getProduct = async (req, res, next) => {
     }) 
 }
 
-const checkErr = (expectedErrors, errors) => {
-    let validationErrors =  [];
-    let expectedErrList = Object.keys(expectedErrors);
-    expectedErrList.forEach(expectedErr => {
-        let error =  errors.find(err => err.param === expectedErr);
-        if (error) {
-            validationErrors.push(error)
-        } 
-    })
-    console.log(validationErrors)
-    if (validationErrors.length > 0) {
-        return validationErrors;
-    } else {
-        return false;
-    } 
-}
 
 exports.addProduct = async (req, res, next) => {
 
     if (!req.file) {
         return res.status(400).json({message: "you should select file"});
     }
-    const {name, descripition, category , sizes, toppings, defaultTopping , specialsAdditions, price} = req.body;
-    let data;
-    let error = validationResult(req).array();
-    if (category == "pizza") {
-        data = {name ,descripition, category, sizes, toppings, defaultTopping , specialsAdditions }
-        let checkErrResult = checkErr(data, error);
-        if(checkErrResult) {
-            await fs.unlinkSync(uploadsPath + req.file.filename)
-            return res.status(400).json({error: checkErrResult});
-        }
-    } else if (category == "drink") {
-        data = {name ,descripition, category, sizes }
-        let checkErrResult = checkErr(data, error);
-        if(checkErrResult) {
-            await fs.unlinkSync(uploadsPath + req.file.filename)
-            return res.status(400).json({error: checkErrResult});
-        }
-    } else if (category == "salads") {
-        data = {name ,descripition, category, sizes, toppings }
-        let checkErrResult = checkErr(data, error);
-        if(checkErrResult) {
-            await fs.unlinkSync(uploadsPath + req.file.filename)
-            return res.status(400).json({error: checkErrResult});
-        }
-    } else if (category == "starters") {
-        data = {name ,descripition, category, sizes }
-        let checkErrResult = checkErr(data, error);
-        if(checkErrResult) {
-            await fs.unlinkSync(uploadsPath + req.file.filename)
-            return res.status(400).json({error: checkErrResult});
-        }
-    } else if (category == "pasta") {
-        data = {name ,descripition, category, price }
-        let checkErrResult = checkErr(data, error);
-        if(checkErrResult) {
-            await fs.unlinkSync(uploadsPath + req.file.filename)
-            return res.status(400).json({error: checkErrResult});
-        }
-    } else if (category == "desert") {
-        data = {name ,descripition, category, price }
-        let checkErrResult = checkErr(data, error);
-        if(checkErrResult) {
-            await fs.unlinkSync(uploadsPath + req.file.filename)
-            return res.status(400).json({error: checkErrResult});
-        }
-    } else {
-        await fs.unlinkSync(uploadsPath + req.file.filename)
-        return res.status(400).json({message : "should select correct category"})
-    }
+    const {name, descripition, category , sizes, toppings, pieces , specialsAdditions, price} = req.body;
+   
     const slug = slugify(name);
     data.slug = slug;
     data.image = req.file.filename;
-    const product = new Product(data)
     try {
+        const product = new Product(data)
         const save = await product.save();
+        const checkProductCategory = await Category.findOne({name: save.category});
+        if (checkProductCategory) {
+            await Category.findByIdAndUpdate(checkProductCategory._id, {$addToSet: {products: save._id}}, {new: false})
+        } else {
+            const newCategoryData = {
+                name: category,
+                products: [save._id]
+            }
+            const newCategory = new Category(newCategoryData);
+            await newCategory.save();
+        }
         res.status(201).json({message: "product is added successfully"})
+        
     } catch (error) {
         await fs.unlinkSync(uploadsPath + req.file.filename)
         res.status(500).json({message: "something went wrong"})
@@ -150,24 +104,27 @@ exports.deleteProduct = async (req, res, next) => {
 }
 exports.updateProduct = async (req, res, next) => {
     const { id , newData } = req.body;
-    console.log(id);
-    const update = await Product.findByIdAndUpdate(id, newData, {new: true});
-    console.log(update);
-    if (update) {
-        return res.status(200).json({
-            id: update.id,
-            name: update.name,
-            category: update.category,
-            image: process.env.URL + update.image,
-            descripition: update.descripition,
-            toppings: update.toppings,
-            defaultTopping: update.defaultTopping,
-            specialsAdditions: update.specialsAdditions,
-            pieces: update.pieces,
-            price: update.price,
-        })
-    } 
-    res.status(500).json({message: "Something went wrong"})
+    if (!id) {
+        return res.status(400).json({message: "ca"})
+    }
+    await Product.findByIdAndUpdate(id, newData, {new: true}, (err, update) => {
+        if (update) {
+            return res.status(200).json({
+                id: update.id,
+                name: update.name,
+                category: update.category,
+                image: process.env.URL + update.image,
+                descripition: update.descripition,
+                toppings: update.toppings,
+                defaultTopping: update.defaultTopping,
+                specialsAdditions: update.specialsAdditions,
+                pieces: update.pieces,
+                price: update.price,
+            })
+        } 
+        res.status(500).json({message: "Something went wrong"})
+    });
+    
 }
 exports.updateProductImage = async (req, res, next) => {
     const image = req.file.filename;
@@ -175,6 +132,10 @@ exports.updateProductImage = async (req, res, next) => {
         return res.status(400).json({message: "you should select an image"})
     }
     const id  = req.body.id; 
+    console.log(req.body.id + " " + image)
+    if (!id) {
+        return res.status(400).json({message: "can't update product with this image"})
+    }
     await Product.findById(id, async (err, doc) => {
         if (err) {
             await fs.unlink(uploadsPath + image, err => {
@@ -186,10 +147,11 @@ exports.updateProductImage = async (req, res, next) => {
         } else {
             try {
                 const update = await Product.findByIdAndUpdate(id, {image}, {new: true})
-                await fs.unlinkSync(uploadsPath + product.image);
+                console.log(update)
+                fs.unlinkSync(uploadsPath + doc.image);
                 res.status(200).json({message: "product image is updated successfully"})
             } catch (error) {
-                res.status(400).json({message: "something went wrong"})
+                res.status(400).json({message: error})
             }
         }
     });
@@ -256,4 +218,57 @@ const product = category => {
         }
 }
 
+*/
+
+
+
+/*
+let data;
+    let error = validationResult(req).array();
+    if (category == "pizza") {
+        data = {name ,descripition, category, sizes, toppings, defaultTopping , specialsAdditions }
+        let checkErrResult = checkErr(data, error);
+        if(checkErrResult) {
+            await fs.unlinkSync(uploadsPath + req.file.filename)
+            return res.status(400).json({error: checkErrResult});
+        }
+    } else if (category == "drink") {
+        data = {name ,descripition, category, sizes }
+        let checkErrResult = checkErr(data, error);
+        if(checkErrResult) {
+            await fs.unlinkSync(uploadsPath + req.file.filename)
+            return res.status(400).json({error: checkErrResult});
+        }
+    } else if (category == "salads") {
+        data = {name ,descripition, category, sizes, toppings }
+        let checkErrResult = checkErr(data, error);
+        if(checkErrResult) {
+            await fs.unlinkSync(uploadsPath + req.file.filename)
+            return res.status(400).json({error: checkErrResult});
+        }
+    } else if (category == "starters") {
+        data = {name ,descripition, category, sizes }
+        let checkErrResult = checkErr(data, error);
+        if(checkErrResult) {
+            await fs.unlinkSync(uploadsPath + req.file.filename)
+            return res.status(400).json({error: checkErrResult});
+        }
+    } else if (category == "pasta") {
+        data = {name ,descripition, category, price }
+        let checkErrResult = checkErr(data, error);
+        if(checkErrResult) {
+            await fs.unlinkSync(uploadsPath + req.file.filename)
+            return res.status(400).json({error: checkErrResult});
+        }
+    } else if (category == "desert") {
+        data = {name ,descripition, category, price }
+        let checkErrResult = checkErr(data, error);
+        if(checkErrResult) {
+            await fs.unlinkSync(uploadsPath + req.file.filename)
+            return res.status(400).json({error: checkErrResult});
+        }
+    } else {
+        await fs.unlinkSync(uploadsPath + req.file.filename)
+        return res.status(400).json({message : "should select correct category"})
+    }
 */
